Refactor some of the tests
It’s a good idea to group tests in a `describe` block for the method
that they’re testing. When we have more tests, this makes it easy to
know which method a test is testing so you can easily find the tests
that you’re looking for.

We can reduce the complexity of these tests and make them easier to
read and understand by using a smaller piece of test data.

This may lead to a few more lines of code but we end up with something
that’s easier to read and understand what it’s testing. It’s also
easier to spot errors since the data set is smaller.