Add WordAnalysis#token_count
We can refactor all of the counting logic into a more generic method
that just counts the occurrences of the items in any Array. Then we can
reuse this counting method for each of the different types of counts
that we want to perform.

Having all of the counting logic in one place makes it easier if we
need to change the way that it works. We now only would need to change
it in once place.

This also makes it easier to figure out the differences between each of
the count methods.