Refactor top methods
Of course, this was not part of requirements but I thought  it would be
a cool thing to show.

If you allow for an optional `count` argument, this method now allows
the user to specify how many results they want to get back, while still
making the usual case easy.