Extra
To begin, I probably wouldn’t actually do this. I just thought this was
a good place to show you how the send method works.

Here, we can pass in the type as an argument. We use that type to
create a string that represents the count method that we need to call
to get the correct set of tokens. For instance, passing in the type
`:letter`, we’re constructing the string “letter_count”.

We’re then passing this string as an argument to the `#send` method.
`#send` is an alternative to using the dot syntax to call a method.

For example:

`self.send(“letter_count”)` is the same as `self.letter_count`. This
lets us use send to be more dynamic about the method that we’re
calling.

Like I said, in this case, this is probably an example of when
refactoring goes too far. I just thought it was a good place to show an
example of how `#send` can be used.